backend/requirements.txt
python-dotenv
llama-index
SQLAlchemy
fastapi
pymupdf
pypdf
pydantic
uvicorn[standard]
python-multipart
